+++ src/InitializerInterface.php
@@ -16,7 +16,7 @@ namespace Zend\ServiceManager;
  *
  * - rename the method `initialize()` to `__invoke()`, and:
  *   - rename the `$serviceLocator` argument to `$container`, and change the
- *     typehint to `Interop\Container\ContainerInterface`
+ *     typehint to `Psr\Container\ContainerInterface`
  *   - swap the order of the arguments (so that `$instance` comes second)
  * - create an `initialize()` method as defined in this interface, and have it
  *   proxy to `__invoke()`, passing the arguments in the new order.