+++ src/Proxy/LazyServiceFactory.php
@@ -7,11 +7,13 @@
 
 namespace Zend\ServiceManager\Proxy;
 
-use Interop\Container\ContainerInterface;
 use ProxyManager\Factory\LazyLoadingValueHolderFactory;
 use ProxyManager\Proxy\LazyLoadingInterface;
+use Psr\Container\ContainerInterface;
 use Zend\ServiceManager\Exception;
 use Zend\ServiceManager\Factory\DelegatorFactoryInterface;
+
+use function sprintf;
 
 /**
  * Delegator factory responsible of instantiating lazy loading value holder proxies of