+++ src/Tool/FactoryCreatorCommand.php
@@ -10,6 +10,11 @@ namespace Zend\ServiceManager\Tool;
 use Zend\ServiceManager\Exception;
 use Zend\Stdlib\ConsoleHelper;
 
+use function array_shift;
+use function class_exists;
+use function in_array;
+use function sprintf;
+
 class FactoryCreatorCommand
 {
     const COMMAND_DUMP = 'dump';
@@ -98,7 +103,7 @@ EOH;
      */
     private function parseArgs(array $args)
     {
-        if (! count($args)) {
+        if (! $args) {
             return $this->createArguments(self::COMMAND_HELP);
         }