+++ test/Factory/InvokableFactoryTest.php
@@ -7,8 +7,8 @@
 
 namespace ZendTest\ServiceManager\Factory;
 
-use Interop\Container\ContainerInterface;
 use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
 use Zend\ServiceManager\Factory\InvokableFactory;
 use ZendTest\ServiceManager\TestAsset\InvokableObject;
 
@@ -25,7 +25,7 @@ class InvokableFactoryTest extends TestCase
 
         $object = $factory($container, InvokableObject::class, ['foo' => 'bar']);
 
-        $this->assertInstanceOf(InvokableObject::class, $object);
-        $this->assertEquals(['foo' => 'bar'], $object->options);
+        self::assertInstanceOf(InvokableObject::class, $object);
+        self::assertEquals(['foo' => 'bar'], $object->options);
     }
 }